requirements.txt
flask==3.1.0
google-cloud-bigquery==3.27.0
supabase==2.11.0
pytest==8.3.4
python-dotenv==1.0.1
gunicorn==23.0.0
